// react.components/src/types/searchTypes.ts
export enum DefaultRequestValue {
  INPUT = '',
  ORIENTATION = 'landscape',
  PER_PAGE = '10',
  PAGE = '1',
}

export enum DefaultSelectValue {
  ORIENTATION = 'Orientation',
  PER_PAGE = 'Per page',
}

export enum LocalStorageRequestValue {
  INPUT = 'inputValue',
  ORIENTATION = 'sortValue',
  PER_PAGE = 'perPAge',
  PAGE = 'page',
}